oppine_back/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
PyJWT
passlib[bcrypt]
bcrypt==4.0.1
sqlalchemy
requests
boto3
pillow
httpx
python-dateutil
pydantic[email]
psycopg2-binary
alembic
sentry-sdk[fastapi]
apscheduler
google-auth
google-auth-oauthlib
google-api-python-client
cryptography